Remove setter method from dataset to fix https://github.com/appliedAI-Initiative/pyDVL/pull/338#discussion_r1306497638

+++ src/pydvl/utils/numeric.py
@@ -189,7 +189,7 @@ def random_powerset_group_conditional(
             rng.shuffle(subset)
             yield subset
         else:
-            yield np.array([])
+            yield np.array([], dtype=int)
 
 
 def random_subset_of_size(s: NDArray[T], size: int) -> NDArray[T]:
